command line argument bug fixed

+++ lib/utils.py
@@ -38,7 +38,7 @@ def parse_args(available_nets):
     parser.add_argument("--resume", type=str, default=None, help="path to model file to be resumed")
     parser.add_argument("--train-net", type=str, default=available_nets[0], help="train on one of following nets: %s"%(str(available_nets)))
     parser.add_argument("--batch-size", type=int, default=8, help="batch size")
-    parser.add_argument("--lr", type=int, default=0.0025, help="learning rate")
+    parser.add_argument("--lr", type=float, default=0.0025, help="learning rate")
     parser.add_argument("--epochs", type=int, default=20, help="run for epochs")
     parser.add_argument("--cache", type=str, default=default_cache, help="where to store training meta-data and final model")
     parser.add_argument("--estimated-3DBB", type=str, default=None, help="use estimated 3DBBs from specified path")
